+++ js/publish/src/index.ts
@@ -7,7 +7,6 @@
  * @module
  */
 export * as Hang from "@moq/hang";
-export * as Json from "@moq/json";
 export * as Net from "@moq/net";
 export * as Signals from "@moq/signals";
 export * as Audio from "./audio";