+++ rs/moq-boy/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.28](https://github.com/moq-dev/moq/compare/moq-boy-v0.2.27...moq-boy-v0.2.28) - 2026-07-18
+
+### Other
+
+- update Cargo.lock dependencies
+
 ## [0.2.27](https://github.com/moq-dev/moq/compare/moq-boy-v0.2.26...moq-boy-v0.2.27) - 2026-07-16
 
 ### Other