+++ rs/moq-relay/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.13.7](https://github.com/moq-dev/moq/compare/moq-relay-v0.13.6...moq-relay-v0.13.7) - 2026-07-18
+
+### Fixed
+
+- *(relay)* avoid websocket teardown panic ([#2390](https://github.com/moq-dev/moq/pull/2390))
+
 ## [0.13.6](https://github.com/moq-dev/moq/compare/moq-relay-v0.13.5...moq-relay-v0.13.6) - 2026-07-16
 
 ### Other