+++ rs/moq-token-cli/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.5.37](https://github.com/moq-dev/moq/compare/moq-token-cli-v0.5.36...moq-token-cli-v0.5.37) - 2026-07-18
+
+### Other
+
+- update Cargo.lock dependencies
+
 ## [0.5.36](https://github.com/moq-dev/moq/compare/moq-token-cli-v0.5.35...moq-token-cli-v0.5.36) - 2026-07-17
 
 ### Other